WS-39: Review markups
