--- frontend/src/components/news-page/NewsCard.scss
@@ -1,11 +1,0 @@
-// .video-container {
-//   width: 100%;
-//   aspect-ratio: 1 / 1;
-//   overflow: hidden;
-//   display: block;
-// }
-
-video {
-  width: 100% !important;
-  height: 100%;
-}
