+++ frontend/src/components/videos-page/VideoCard.scss
@@ -1,0 +1,4 @@
+video.reddit-whale-video {
+  width: 100%;
+  height: 100%;
+}
